+++ app/src/main/AndroidManifest.xml
@@ -17,6 +17,7 @@
         <activity
             android:name=".core.RootActivity"
             android:exported="true"
+            android:configChanges="orientation"
             android:label="@string/app_name"
             android:theme="@style/Theme.TMDB.NoActionBar">
             <intent-filter>
